Added roll_now support to Roll and Throw
By default, roll_now is True for Roll and Throw, but False for Die. Die
objects are most often instantiated as part of building a Roll, which
will roll all member Die objects in a Throw. Meanwhile, a Roll object
itself may be (re-)constructed from old Die objects, in which case we
would prefer not to scramble the Die objects.